fix: corrige os relacionamentos e atributos das entidades

+++ src/main/java/api/carometro/repositories/HistoricoProfissionalRepository.java
@@ -3,5 +3,5 @@ package api.carometro.repositories;
 import api.carometro.models.HistoricoProfissional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoricoProfissionalRepository extends JpaRepository<HistoricoProfissional, Integer> {
+public interface HistoricoProfissionalRepository extends JpaRepository<HistoricoProfissional, Long> {
 }
